dict.cpp
The program takes in a text file (in this case, the first few chapters
of "The Adventures of Huckleberry Finn") and prompts the user to input a
word, phrase or sentence to find within the text. The program outputs
the 10 most identical word/phrase/sentence by calculating the hamming
distance of the input and each and every word/phrase/sentence.  In my
case, I used a set to alphabetically store the parsed words, phrases and
sentences so the program can find the correct outputs quickly.

data.txt contains our input text file. Sample-output.txt contains
examples to what my program will output based on the user input.