STL
STL implementations with Linked list and Vector List. Implemented
functions that inserted at the front and back of the list, list
reversal, list comparison and such. At the end, our program would check
how much memory was allocated and deallocated to see whether or not we
had a memory leak. README for more info